# agent/openapi.json
{
  "components": {
    "schemas": {
...
      "ArpPollRequest": {
        "properties": {
          "community": {
            "format": "password",
            "title": "Community",
            "type": "string",
            "writeOnly": true
          },
          "router_ip": {
            "title": "Router Ip",
            "type": "string"
          },
          "timeout_ms": {
            "default": 10000,
            "maximum": 120000.0,
            "minimum": 500.0,
            "title": "Timeout Ms",
            "type": "integer"
          },
          "version": {
            "default": "2c",
            "enum": [
              "1",
              "2c",
              "3"
            ],
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "router_ip",
          "community"
        ],
        "title": "ArpPollRequest",
        "type": "object"
      },
      "ArpPollResponse": {
        "properties": {
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ArpEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "raw_lines": {
            "items": {
              "type": "string"
            },
            "title": "Raw Lines",
            "type": "array"
          }
        },
        "required": [
          "entries",
          "raw_lines",
          "duration_ms"
        ],
        "title": "ArpPollResponse",
        "type": "object"
      },
      "DnsBatchEntry": {
        "properties": {
          "ip": {
            "title": "Ip",
            "type": "string"
          },
          "resolution": {
            "$ref": "#/components/schemas/DnsResolution"
          }
        },
        "required": [
          "ip",
          "resolution"
        ],
        "title": "DnsBatchEntry",
        "type": "object"
      },
      "DnsBatchRequest": {
        "properties": {
          "concurrency": {
            "default": 16,
            "maximum": 64.0,
            "minimum": 1.0,
            "title": "Concurrency",
            "type": "integer"
          },
          "dns_server": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dns Server"
          },
          "ips": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Ips",
            "type": "array"
          }
        },
        "required": [
          "ips"
        ],
        "title": "DnsBatchRequest",
        "type": "object"
      },
      "DnsForwardRequest": {
        "properties": {
          "dns_server": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dns Server"
          },
          "hostname": {
            "title": "Hostname",
            "type": "string"
          },
          "timeout_ms": {
            "default": 2500,
            "maximum": 30000.0,
            "minimum": 200.0,
            "title": "Timeout Ms",
            "type": "integer"
          }
        },
        "required": [
          "hostname"
        ],
        "title": "DnsForwardRequest",
        "type": "object"
      },
      "DnsResolution": {
        "properties": {
          "forward": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forward"
          },
          "reverse": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reverse"
          }
        },
        "title": "DnsResolution",
        "type": "object"
      },
      "DnsReverseRequest": {
        "properties": {
          "dns_server": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dns Server"
          },
          "ip": {
            "title": "Ip",
            "type": "string"
          },
          "timeout_ms": {
            "default": 2500,
            "maximum": 30000.0,
            "minimum": 200.0,
            "title": "Timeout Ms",
            "type": "integer"
          }
        },
        "required": [
          "ip"
        ],
        "title": "DnsReverseRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NmapDiscoverHostsRequest": {
        "properties": {
          "target": {
            "title": "Target",
            "type": "string"
          },
          "timeout_ms": {
            "default": 90000,
            "maximum": 900000.0,
            "minimum": 1000.0,
            "title": "Timeout Ms",
            "type": "integer"
          }
        },
        "required": [
          "target"
        ],
        "title": "NmapDiscoverHostsRequest",
        "type": "object"
      },
...
      "NmapResult": {
        "properties": {
          "alive": {
            "title": "Alive",
            "type": "boolean"
          },
          "ip": {
            "title": "Ip",
            "type": "string"
          },
          "mac": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mac"
          },
          "os": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Os"
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/NmapPort"
            },
            "title": "Ports",
            "type": "array"
          }
        },
        "required": [
          "ip",
          "alive"
        ],
        "title": "NmapResult",
        "type": "object"
      },
...
      "SnmpRoutesRequest": {
        "properties": {
          "community": {
            "format": "password",
            "title": "Community",
            "type": "string",
            "writeOnly": true
          },
          "router_ip": {
            "title": "Router Ip",
            "type": "string"
          },
          "timeout_ms": {
            "default": 15000,
            "maximum": 120000.0,
            "minimum": 500.0,
            "title": "Timeout Ms",
            "type": "integer"
          },
          "version": {
            "default": "2c",
            "enum": [
              "1",
              "2c",
              "3"
            ],
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "router_ip",
          "community"
        ],
        "title": "SnmpRoutesRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DA-INVENT Agent",
    "version": "0.2.4"
  },
  "openapi": "3.1.0",
  "paths": {
    "/exec/arp-poll": {
      "post": {
        "operationId": "exec_arp_poll_exec_arp_poll_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArpPollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArpPollResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exec Arp Poll"
      }
    },
    "/exec/dns-batch": {
      "post": {
        "operationId": "exec_dns_batch_exec_dns_batch_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DnsBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DnsBatchEntry"
                  },
                  "title": "Response Exec Dns Batch Exec Dns Batch Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exec Dns Batch"
      }
    },
    "/exec/dns-forward": {
      "post": {
        "operationId": "exec_dns_forward_exec_dns_forward_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DnsForwardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Exec Dns Forward Exec Dns Forward Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exec Dns Forward"
      }
    },
    "/exec/dns-reverse": {
      "post": {
        "operationId": "exec_dns_reverse_exec_dns_reverse_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DnsReverseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsResolution"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exec Dns Reverse"
      }
    },
    "/exec/nmap-discover": {
      "post": {
        "operationId": "exec_nmap_discover_exec_nmap_discover_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NmapDiscoverHostsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NmapResult"
                  },
                  "title": "Response Exec Nmap Discover Exec Nmap Discover Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exec Nmap Discover"
      }
    },
    "/exec/nmap-port-scan": {
      "post": {
        "operationId": "exec_nmap_port_scan_exec_nmap_port_scan_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NmapPortScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/NmapResult"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Exec Nmap Port Scan Exec Nmap Port Scan Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exec Nmap Port Scan"
      }
    },
...
    "/exec/snmp-routes": {
      "post": {
        "operationId": "exec_snmp_routes_exec_snmp_routes_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnmpRoutesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SnmpRoute"
                  },
                  "title": "Response Exec Snmp Routes Exec Snmp Routes Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exec Snmp Routes"
      }
    },
    "/exec/snmp-walk": {
      "post": {
        "operationId": "exec_snmp_walk_exec_snmp_walk_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnmpWalkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SnmpWalkRow"
                  },
                  "title": "Response Exec Snmp Walk Exec Snmp Walk Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exec Snmp Walk"
      }
    },
    "/exec/ssh-exec": {
      "post": {
        "operationId": "exec_ssh_run_exec_ssh_exec_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SshExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exec Ssh Run"
      }
    },
    "/exec/winrm-exec": {
      "post": {
        "operationId": "exec_winrm_run_exec_winrm_exec_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WinrmExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WinrmExecResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exec Winrm Run"
      }
    },
...
    "/whoami": {
      "get": {
        "operationId": "whoami_whoami_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhoamiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Whoami"
      }
    }
  }
}